pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gcp-memory-bank"
version = "0.1.0"
description = "Production-grade Python SDK for Google Gemini Enterprise Agent Platform Memory Bank"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
authors = [
    {name = "Hermes Agent", email = "hermes@local"},
]
keywords = ["google-cloud", "memory-bank", "agent-memory", "vertex-ai", "gemini"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "google-cloud-aiplatform[agent_engines]>=1.90.0",
    "google-genai>=1.0.0",
    "pydantic>=2.0.0",
    "tenacity>=9.0.0",
    "jinja2>=3.1.0",
    "aiohttp>=3.9.0",
    "structlog>=24.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.6.0",
    "mypy>=1.10.0",
    "pre-commit>=3.7.0",
]

[project.urls]
Homepage = "https://github.com/hermes/gcp-memory-bank"
Documentation = "https://docs.cloud.google.com/gemini-enterprise-agent-platform/scale/memory-bank"

[tool.hatch.build.targets.wheel]
packages = ["src/memory_bank"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
